+++ server/src/main/java/org/opensearch/action/bulk/BulkRequest.java
@@ -38,6 +38,7 @@ import org.opensearch.Version;
 import org.opensearch.action.ActionRequest;
 import org.opensearch.action.ActionRequestValidationException;
 import org.opensearch.action.CompositeIndicesRequest;
+import org.opensearch.action.DocRequest;
 import org.opensearch.action.DocWriteRequest;
 import org.opensearch.action.delete.DeleteRequest;
 import org.opensearch.action.index.IndexRequest;
@@ -111,6 +112,7 @@ public class BulkRequest extends ActionRequest implements CompositeIndicesReques
         if (in.getVersion().onOrAfter(Version.V_2_14_0) && in.getVersion().before(Version.V_3_0_0)) {
             in.readInt(); // formerly batch_size
         }
+        requests.stream().map(DocRequest::index).forEach(indices::add);
     }
 
     public BulkRequest(@Nullable String globalIndex) {